[TASK] Simplify setup command & remove dead code
